refactor(schedule): address now-view review comments
Colocates the Now board in its route file, switches it to suspense
queries with a sets loader prefetch, and rewrites the classifier
comments as what/why JSDoc.
