+++ src/routes/festivals/$festivalSlug/editions/$editionSlug/schedule/now.tsx
@@ -1,7 +1,24 @@
-import { createFileRoute, redirect } from "@tanstack/react-router";
-import { ScheduleTabNow } from "@/pages/EditionView/tabs/ScheduleTab/now/NowBoard";
+import { useSuspenseQuery } from "@tanstack/react-query";
+import {
+  createFileRoute,
+  Link,
+  redirect,
+  useParams,
+} from "@tanstack/react-router";
+import { StageBadge } from "@/components/StageBadge";
+import { setsByEditionQuery } from "@/api/sets/useSetsByEdition";
+import { stagesByEditionQuery } from "@/api/stages/useStagesByEdition";
+import { useFestivalEdition } from "@/contexts/FestivalEditionContext";
 import { getFestivalPhase } from "@/lib/festivalPhase";
+import {
+  classifyNowNextByStage,
+  type NowNextClassification,
+} from "@/lib/nowNext";
 import { canShowTime } from "@/lib/scheduleReveal";
+import { sortStagesByOrder } from "@/lib/stageUtils";
+import { formatTimeOnly } from "@/lib/timeUtils";
+import type { FestivalSet } from "@/api/sets/types";
+import type { Stage } from "@/api/stages/types";
 
 export const Route = createFileRoute(
   "/festivals/$festivalSlug/editions/$editionSlug/schedule/now",
@@ -27,4 +44,106 @@ export const Route = createFileRoute(
       });
     }
   },
+  loader: ({ context }) => {
+    void context.queryClient.ensureQueryData(
+      setsByEditionQuery(context.edition.id),
+    );
+  },
 });
+
+function ScheduleTabNow() {
+  const { festival } = useFestivalEdition();
+  const { edition } = Route.useRouteContext();
+  const { data: sets } = useSuspenseQuery(setsByEditionQuery(edition.id));
+  const { data: stages } = useSuspenseQuery(stagesByEditionQuery(edition.id));
+
+  const byStage = classifyNowNextByStage(sets, new Date());
+  const rows = sortStagesByOrder([...stages]).flatMap((stage) => {
+    const classification = byStage.get(stage.id);
+    if (!classification) return [];
+    if (!classification.nowPlaying.length && !classification.next.length) {
+      return [];
+    }
+    return [{ stage, classification }];
+  });
+
+  if (!rows.length) {
+    return (
+      <div className="text-center text-purple-300 py-12">
+        <p>Nothing on right now — see the timeline for the full schedule.</p>
+      </div>
+    );
+  }
+
+  return (
+    <div className="space-y-3">
+      {rows.map(({ stage, classification }) => (
+        <NowStageRow
+          key={stage.id}
+          stage={stage}
+          classification={classification}
+          timezone={festival.timezone}
+        />
+      ))}
+    </div>
+  );
+}
+
+interface NowStageRowProps {
+  stage: Stage;
+  classification: NowNextClassification<FestivalSet>;
+  timezone: string;
+}
+
+function NowStageRow({ stage, classification, timezone }: NowStageRowProps) {
+  return (
+    <div className="bg-white/10 backdrop-blur-md border border-purple-400/30 rounded-lg p-4 space-y-2">
+      <StageBadge
+        stageName={stage.name}
+        stageColor={stage.color ?? undefined}
+        size="sm"
+      />
+
+      {classification.nowPlaying.map((set) => (
+        <SetLine key={set.id} set={set} timezone={timezone} live />
+      ))}
+      {classification.next.map((set) => (
+        <SetLine key={set.id} set={set} timezone={timezone} />
+      ))}
+    </div>
+  );
+}
+
+interface SetLineProps {
+  set: FestivalSet;
+  timezone: string;
+  live?: boolean;
+}
+
+function SetLine({ set, timezone, live = false }: SetLineProps) {
+  const { festivalSlug, editionSlug } = useParams({
+    from: "/festivals/$festivalSlug/editions/$editionSlug",
+  });
+
+  const time = live
+    ? `until ${formatTimeOnly(set.time_end, null, true, timezone)}`
+    : `at ${formatTimeOnly(set.time_start, null, true, timezone)}`;
+
+  return (
+    <div className="flex items-center gap-2 text-sm">
+      {live ? (
+        <span className="h-2 w-2 shrink-0 rounded-full bg-red-500 animate-pulse" />
+      ) : (
+        <span className="shrink-0 text-purple-300">Next:</span>
+      )}
+      <Link
+        to="/festivals/$festivalSlug/editions/$editionSlug/sets/$setSlug"
+        params={{ festivalSlug, editionSlug, setSlug: set.slug }}
+        className="text-white font-medium hover:text-purple-300 transition-colors truncate"
+      >
+        {set.name}
+      </Link>
+      <span className="ml-auto shrink-0 text-purple-200">{time}</span>
+    </div>
+  );
+}
